refactor: split utils
